Rename `azul.plugins.metadata.hca.stages` to `.service`
